requirements.txt
-e git+git://github.com/django/django.git@1.6b2#egg=django
psycopg2==2.5.1
redis==2.8.0
django-celery==3.0.23
django-autocomplete-light==1.4.1
django-model-utils==1.5.0
